refactor: move exclusion logic to a helper
